# [perf] use uv for venv creation and pip install