… sky/templates/kubernetes-ray.yml.j2
… sky/utils/controller_utils.py