<!-- Describe the changes in this PR -->
This was previously merged in as #4394, which was reverted in #4401 due to a bug.
This PR:
- readds the old uv changes
- fixes the bug on Azure (see below)
- uses the `--seed` flag with `uv venv` to make sure pip is installed in the venv

**Performance impact:**
- Cluster provisioning:
  - Using skypilot custom images:
    - AWS: saves about 0.5s (8.5s -> 8s)
    - Azure: saves about 5s (26s -> 21s)
  - Using stock ubuntu images **~3x speedup**:
    - AWS: saves about 30s (45s -> 15s)
    - Azure: saves about 100s (130s -> 30s)
    - These results likely also apply to other user-provided images, including custom kubernetes images.
- Job controller setup/cloud dependency installation saves about 45s (85s -> 25s) on first run
  - with GCP, AWS, and Azure enabled
  - on run 2+ with the controller already setup, saves about 3s (6.5s -> 3.5s)

Azure bug:
`uv pip install skypilot[azure]` would fail to install azure-cli and complain as follows:
```
  × No solution found when resolving dependencies:
  ╰─▶ Because there is no version of azure-keyvault-administration==4.4.0b2
      and azure-cli>=2.65.0 depends on azure-keyvault-administration==4.4.0b2,
      we can conclude that azure-cli>=2.65.0 cannot be used.
      And because only the following versions of azure-cli are available:
          azure-cli<=2.65.0
          azure-cli==2.66.0
          azure-cli==2.67.0
      and skypilot[azure]==1.0.0.dev0 depends on azure-cli>=2.65.0, we can
      conclude that skypilot[azure]==1.0.0.dev0 cannot be used.
      And because only skypilot[azure]==1.0.0.dev0 is available and you
      require skypilot[azure], we can conclude that your requirements are
      unsatisfiable.

      hint: azure-keyvault-administration was requested with a pre-release
      marker (e.g., azure-keyvault-administration==4.4.0b2), but pre-releases
      weren't enabled (try: `--prerelease=allow`)
```
This is because azure-cli directly depends on a pre-release version of `azure-keyvault-administration` (and several other packages). Normal pip handles this by allowing the pre-release version since it's specified as an explicit version requirement, however `uv` by default does not consider pre-releases at all. See https://docs.astral.sh/uv/pip/compatibility/#pre-release-compatibility.
We don't want to use `uv pip install --prerelease=allow skypilot[azure]` because this will allow pre-release versions to satisfy _any_ dependency, not just these explicitly specified ones. I tested this and the resolved version was different for a few packages. This seems like an unacceptable difference, especially since we will be unnecessarily relying on pre-release packages.
Once azure-cli is installed, `uv pip install skypilot[azure]` will work just fine, since it doesn't need to worry about resolving azure-cli's dependencies. So, we just need to find a way to install azure-cli. `pip install azure-cli` works, but it's VERY slow (takes >60s on my machine). `uv pip install --prerelease=allow azure-cli` is much faster, so I went with this.
Technically, `uv pip install --prerelease=allow azure-cli` uses different resolution than `pip install azure-cli`, for the same reason described above for installing `skypilot[azure]`. However, I tested this, and in practice only one package has a different version. (azure-mgmt-search 9.1.0 -> 9.2.0b2)

<!-- Describe the tests ran -->
<!-- Unit tests (tests/test_*.py) are part of GitHub CI; below are tests that launch on the cloud. -->

Tested (run the relevant ones):

- [x] Code formatting: `bash format.sh`
- [ ] Any manual or new tests for this PR (please specify below)
- [ ] All smoke tests: `pytest tests/test_smoke.py` 
- [ ] Relevant individual smoke tests: `pytest tests/test_smoke.py::test_fill_in_the_name` 
- Backward compatibility tests: `conda deactivate; bash -i tests/backward_compatibility_tests.sh`
  - [x] k8s
  - [x] aws
  - [x] azure 
    - technically I can't get this to pass - everything passes except step 7 which fails because it takes 50 minutes to schedule the job lol