@cg505 2024-11-25
@romilbhardwaj Rather than conditionally using pip/uv based on what's installed, I just changed the RAY_INSTALLATION_COMMANDS to also install uv if it's missing. I think this is better because we need different behavior between pip and uv, so trying to make things compatible with both is going to get increasingly difficult. As a brief example, `pip uninstall` needs the `-y` flag to work non-interactively, but this flag does not exist in `uv pip uninstall`.