@cg505 2024-12-03
> Can we have a quick number for the performance improvement in the PR description, for future reference?

@Michaelvll updated PR description!